fix: Destroy authjs session using payload logout hook instead of overriding logout endpoints
Until now, I have overwritten the logout endpoint to delete the authjs session cookies.
However, this could also be achieved by adding an afterLogout hook and setting the cookies with Next.js.
This simplifies the process and is likely to be the correct approach.

This also corrects the logout server action that was introduced with payloadcms/payload#11900.